Refactor: Make code cleaner
