+++ app/src/main/java/com/example/buibinhminh/ui/playlist/PlaylistGridView.kt
@@ -1,27 +1,30 @@
-package com.example.buibinhminh.ui.musicApp
+package com.example.buibinhminh.ui.playlist
 
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.itemsIndexed
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.example.buibinhminh.data.Song
 
 @Composable
-fun PlaylistListView(
+fun PlaylistGrid(
     songs: List<Song>,
     onDeleteClick: (Song) -> Unit,
     modifier: Modifier = Modifier
 ) {
-    LazyColumn(
-        modifier = modifier.padding(horizontal = 16.dp)
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(2),
+        verticalArrangement = Arrangement.spacedBy(0.dp),
+        modifier = modifier
     ) {
         itemsIndexed(
             songs,
             key = { index , song -> song.id }
         ) { index, song ->
-            SongListItem(
+            SongGridItem(
                 song = song,
                 onDeleteClick = onDeleteClick
             )
