+++ app/src/main/java/com/example/buibinhminh/ui/playlist/PlaylistListView.kt
@@ -1,30 +1,27 @@
-package com.example.buibinhminh.ui.musicApp
+package com.example.buibinhminh.ui.playlist
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.itemsIndexed
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.example.buibinhminh.data.Song
 
 @Composable
-fun PlaylistGrid(
+fun PlaylistListView(
     songs: List<Song>,
     onDeleteClick: (Song) -> Unit,
     modifier: Modifier = Modifier
 ) {
-    LazyVerticalGrid(
-        columns = GridCells.Fixed(2),
-        verticalArrangement = Arrangement.spacedBy(0.dp),
-        modifier = modifier
+    LazyColumn(
+        modifier = modifier.padding(horizontal = 16.dp)
     ) {
         itemsIndexed(
             songs,
             key = { index , song -> song.id }
         ) { index, song ->
-            SongGridItem(
+            SongListItem(
                 song = song,
                 onDeleteClick = onDeleteClick
             )
